+++ packages/text-input/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [3.4.0] - 2019-07-30 - [HUB-297](https://creditandfinance.atlassian.net/browse/HUB-297)
+### Added
+- Add 'xs' size to text-input.
+- Fixed disabled state still get hover bug.
+
 ## [3.3.0] - 2019-06-27 - [HUB-257](https://creditandfinance.atlassian.net/browse/HUB-257)
 ### Changed
 - Allow more types to input like email, tel, number. 'text' is default.